Screening Round - Coding Exercise
Problem: Develop a single page mobile app using Flutter (page layout as shown in the image below) and compile as well as run it on both Android & iOS emulators, or devices. Submit all the artifacts to complete your first round of interview that are mentioned at the end of this document.

Requirement Description:
● Widget 2 (as shown in Image 1.1) should be tappable
● On each tapping on Widget 2, the Widget 3 should display a random no between 0 - 59
KJBN Labs Pvt Ltd
 1
● If the random no displayed on Widget 3, equals to the ‘seconds’ part of current timestamp, then widget 1 should display a success message
● If the random no displayed on Widget 3, not equals to the seconds part of current timestamp, then widget 1 should display a try again message
● The success message should also display, the total no of times the success was achieved since the user installed the app
Following are the conditions.
● The app must use FlutterBloc / or any of the state management library/ or change notifier
APIs for its state management
● The app must use Repository Provider for storing and retrieving data
The look and feel of the app is up to the developer, however a good user experience will be given extra perks.
The developer must use the latest stable version of Flutter, and develop the app from scratch.
Coding standards followed will also be given extra perks.
The candidate is expected to give a creative and suitable name to the app, and that is up to the candidate’s creativity, and the candidate should be able to reason out why that name was chosen.
Artifacts to be submitted are as below (on orrbbefeofroer2e9t7htOhcOt 2c0t2320) 23 6:00 PM IST):
1. Github repository link of the source code
a. The github repo should contain the Readme.md file with all the details &
documentation
b. The readme.md should also contain the details of the developer/candidate
2. Two screen capture videos ( Android & iOS ) that shows the recording of the app running and the functionalities used ( on an emulator or real device )
Please email us at help@thetriggerapp.com